[#116] refactor: 카테고리 쿼리 패키지 구조 수정
